fix:add validation to hash command
