+++ src/fs/delete.js
@@ -1,6 +1,6 @@
 import { unlink } from 'fs/promises';
 import { throwOperationFailed } from '../validation.js';
-import { isExists } from './access.js';
+import { isExists } from '../fs-helpers.js';
 
 export const remove = async (fileToRemove) => {
     const isFileToRemoveExists = await isExists(fileToRemove);
