+++ src/fs/read.js
@@ -1,7 +1,7 @@
 import { createReadStream } from 'fs';
 import { stdout } from 'process';
 import { throwOperationFailed } from '../validation.js';
-import { isExists } from './access.js';
+import { isExists } from '../fs-helpers.js';
 import { printCurrentDirectory } from '../helpers.js';
 
 export const read = async (fileName) => {
